@@ -4,7 +4,7 @@ homepage: https://github.com/yoktavian/time-picker-sheet
 version: 0.0.2+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.3.0 <4.0.0"
 
 dependencies:
   flutter:
