@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:time_picker_sheet/widget/provider/time_picker.dart';
 
 class SheetHeader extends StatelessWidget {
-  const SheetHeader({Key? key}) : super(key: key);
+  const SheetHeader({super.key});
 
   @override
   Widget build(BuildContext context) {
